レスポンスに関するコードをviewsに移動
